Always use the pm-utils backend, even if logind is running
debian/patches/always_use_pm-utils_backend.patch: Always use the pm-utils
backend, even if logind is running. The availability of logind does not
necessarily imply, that systemd is running (in a recent enough version),
which is required for a successful suspend/hibernate.

Once we have v204 in unstable and systemd-shim for sysvinit we can
revert this change again.